+++ src/Controller/EspeceController.php
@@ -18,6 +18,7 @@ class EspeceController extends AbstractController
     {
         return $this->render('espece/index.html.twig', [
             'especes' => $especeRepository->findAll(),
+            'nbEspeces' => sizeof($especeRepository->findAll()),
         ]);
     }
 
@@ -65,10 +66,21 @@ class EspeceController extends AbstractController
 
     public function delete(Request $request, Espece $espece): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$espece->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($espece);
-            $entityManager->flush();
+        if (count($espece->getAnimals()) === 0) {
+            if ($this->isCsrfTokenValid('delete'.$espece->getId(), $request->request->get('_token'))) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($espece);
+                $entityManager->flush();
+                $this->addFlash(
+                'warning',
+                'Espece supprimée...'
+            );
+            }
+        } else {
+            $this->addFlash(
+                'warning',
+                'Cette espèce n\'est pas encore en voie de disparition !'
+            );
         }
 
         return $this->redirectToRoute('espece_index');
